refactor(react): use style under root to build lib/style
