+++ packages/react/package.json
@@ -21,7 +21,7 @@
     "start": "webpack-dev-server --inline --config ./config/webpack.dev.config.js",
     "build": "webpack --config ./config/webpack.prod.config.js",
     "test": "jest --config ./jest.config.js",
-    "compile": "mv lib/style . && rm -rf lib && cross-env NODE_ENV=production babel src --out-dir lib --copy-files && mv style/ lib/ && ../../node_modules/less/bin/lessc lib/style/camera-view.less lib/style/camera-view.css"
+    "compile": "rm -rf lib && cross-env NODE_ENV=production babel src --out-dir lib --copy-files && ../../node_modules/less/bin/lessc style/camera-view.less lib/style/camera-view.css && cp -r style/images lib/style/"
   },
   "bugs": {
     "url": "https://github.com/Skywatch24/JS-Library/issues"
