+++ packages/react/src/CameraView.js
@@ -10,7 +10,7 @@ import {Requests} from '@skywatch/api';
 import ArchivesPlayer from './ArchivesPlayer';
 import FlvPlayer from './FlvPlayer';
 import {useInterval, usePageVisibility} from './hooks';
-import LoadingSpinner from '../lib/style/images/loading.gif';
+import LoadingSpinner from '../style/images/loading.gif';
 import {STATUS_OK, STATUS_END, STATUS_HOLE} from './Constants';
 
 const hide_ff = false;
